+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+### Fixed
+- Trying to access array offset on value of type bool
+  [#54](https://github.com/gino0631/nextcloud-metadata/issues/54)
 
 ## 0.11.1 – 2020-01-21
 ### Fixed