+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Support for image Subject metadata written by Windows
+  [#57](https://github.com/gino0631/nextcloud-metadata/issues/57)
+
 ### Fixed
 - Trying to access array offset on value of type bool
   [#54](https://github.com/gino0631/nextcloud-metadata/issues/54)