+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   [#41](https://github.com/gino0631/nextcloud-metadata/issues/41)
 - Support for image Subject metadata written by Windows
   [#57](https://github.com/gino0631/nextcloud-metadata/issues/57)
+- Support for ratings
+  [#39](https://github.com/gino0631/nextcloud-metadata/issues/39)
 - Translations to many languages from
   [Transifex](https://www.transifex.com/nextcloud/nextcloud/metadata/)